backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic[email]==2.5.2
pydantic-settings==2.1.0
celery[redis]==5.3.6
redis==5.0.1
boto3==1.34.0
python-dateutil==2.8.2
websockets==12.0
aiofiles==23.2.1
email-validator==2.1.0
